Allow package manager to fully control install/uninstall steps
